Extracted common code to a base class for handler tests
